Update MultiKueue integration tests
